rework product subscription
